Use fixtures by reference to test couch schemas
